Move the mapper's split layout somewhere the table page can reach it
The resizable two-column layout was private to the mapper's editor. Nothing about
it is mapper-specific, and the custom table editor wants the same arrangement, so
it moves to components/split-layout.tsx unchanged.
